requirements.txt
matplotlib==3.7.1
numpy==1.23.5
Pillow==9.4.0
Pillow==9.5.0
scikit_learn==1.2.2
torch==2.0.0
torchmetrics==0.11.2
torchvision==0.15.0
